requirements.txt
schema>=0.7.4
requests>=2.25.0
requests-futures>=1.0.0
tqdm>=4.60.0
boto3>=1.18.0
fire>=0.4.0
orjson>=3.6.4
smart_open>=5.2.1
python-dateutil>=2.8.2
mgzip==0.2.1
pandas>=1.2.0
lz4>=3.1.3
Pillow>=7.0.0
google-cloud-storage==2.7.0
scikit-image==0.20.0